FYP-69: Environment Wrapper Test (#34)
* fix type of reset_pos

* add environment tests

* add python test workflow

* add missing requirmeents

* fix style issues

* ignore stupid rule

* fix python 3.10

* use 3.10.11

+++ 04_Repository/Gym/bullet_drone_env.py
@@ -81,4 +81,4 @@ class BulletDroneEnv(gym.Env):
 
         reset_pos = self.goal_state + np.array([x_offset, 0, y_offset], dtype=np.float32)
 
-        return reset_pos
+        return reset_pos.astype(np.float32)
